+++ server/src/main/java/ru/practicum/shareit/item/api/repository/ItemRepository.java
@@ -11,9 +11,9 @@ import ru.practicum.shareit.request.entity.ItemRequest;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
     @Query("select i from Item i where i.owner.id = ?1")
-    List<Item> findAllByOwner_Id(Integer id, Pageable pageable);
+    List<Item> findAllByOwner_Id(Long id, Pageable pageable);
 
     @Query("select i from Item i " +
             "where ( " +
@@ -24,12 +24,12 @@ public interface ItemRepository extends JpaRepository<Item, Integer> {
             @Param("search") String text, Pageable pageable);
 
     @Query("select not(count(i) > 0) from Item i where i.id = ?1 and i.owner.id = ?2")
-    boolean notExistsByIdAndOwner_Id(int itemId, Integer ownerId);
+    boolean notExistsByIdAndOwner_Id(Long itemId, Long ownerId);
 
     @Transactional
     @Modifying
     @Query("delete from Item i where i.id = ?1 and i.owner.id = ?2")
-    void deleteByIdAndOwner_Id(Integer itemId, Integer ownerId);
+    void deleteByIdAndOwner_Id(Long itemId, Long ownerId);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
@@ -37,7 +37,7 @@ public interface ItemRepository extends JpaRepository<Item, Integer> {
     void updateNameAndDescriptionById(
             @Param("name") String name,
             @Param("description") String description,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
@@ -45,7 +45,7 @@ public interface ItemRepository extends JpaRepository<Item, Integer> {
     void updateNameAndAvailableById(
             @Param("name") String name,
             @Param("available") boolean available,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
@@ -53,31 +53,31 @@ public interface ItemRepository extends JpaRepository<Item, Integer> {
     void updateDescriptionAndAvailableById(
             @Param("description") String description,
             @Param("available") boolean available,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("update Item i set i.name = :name where i.id = :id")
     void updateNameById(
             @Param("name") String name,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("update Item i set i.description = :description where i.id = :id")
     void updateDescriptionById(
             @Param("description") String description,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Transactional
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("update Item i set i.available = :available where i.id = :id")
     void updateAvailableById(
             @Param("available") boolean available,
-            @Param("id") int id);
+            @Param("id") Long id);
 
     @Query("select i from Item i where i.request.id = ?1")
-    List<Item> getByRequest_Id(Integer id);
+    List<Item> getByRequest_Id(Long id);
 
     @Query("select i from Item i where i.request in ?1")
     List<Item> findByRequestIn(List<ItemRequest> requests);
